services/auth/requirements.txt
fastapi==0.115.0
pydantic==2.9.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6